Fixed most stuff
Still missing custom layerDepth based on the index that the grid is in, in the grids ing GridManager. Also the Agent need to move when the tile is a collider. Make a button for it and the other choices. Make the agent move if the tile isWalkable and if there is a collider on, then it cant move there.